More things in tests.

+++ src/include/TempLat/util/ndloop.h
@@ -36,17 +36,17 @@ namespace TempLat
   void NDLoop(const View &view, const Functor &functor)
   {
     std::array<ptrdiff_t, NDim> extents;
-    for (ptrdiff_t i = 0; i < NDim; ++i)
+    for (size_t i = 0; i < NDim; ++i)
       extents[i] = view.extent(i);
 
     std::array<ptrdiff_t, NDim> idx;
-    for (ptrdiff_t i = 0; i < NDim; ++i)
+    for (size_t i = 0; i < NDim; ++i)
       idx[i] = 0;
 
     while (true) {
       std::apply(functor, idx);
-      ptrdiff_t dim = NDim - 1;
-      while (dim < NDim) {
+      ptrdiff_t dim = (ptrdiff_t)NDim - 1;
+      while (dim < (ptrdiff_t)NDim) {
         idx[dim]++;
         if (idx[dim] < extents[dim]) {
           break;
